requirements.txt
python-telegram-bot==20.7
playwright==1.28.0
pyee==9.0.4
requests==2.31.0